fix(landing): replace the navbar with the logo
This working ok when the user is signed out. However, the navbar view
changes when the user is signed in - and this was distoring the layout
of the other elements on the page.

In any case, we only needed the logo on this page for aesthetics and
this is what the navbar was meant to do.