fix(styles): apply opacity on non-touch devices only
This ensures the reduced-opacity-on-hover feauture on the category
section of the profile page is enabled only for non-touch devices. It's
important to do this so that the opacity isn't reduced for touch devices
as hovering isn't available to them.

context: the opacity is restored when you hover over a certain category.